ANDROID: intel: Do not use libpciaccess on Android
This patch makes the code not rely anymore on libpciaccess when compiled
for Android to eliminate ioperm() and iopl() syscalls required by that
library. As a side effect, the mappable aperture size is hardcoded to 64
MiB on Android, however nothing seems to rely on this value anyway, as
checked be grepping relevant code in drm_gralloc and Mesa.

Change-Id: I306229b1e5d1500ad72ee1684ea0b03b4db3efc9
Signed-off-by: Tomasz Figa <tfiga@google.com>
(cherry picked from commit f6a1130dffae8de9ddd0c379066daf1df27fc8af)